+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## 0.0.1-dev-17-feature/DHP-23-field-type-names
+## 0.0.1-dev-18-feature/DHP-23-field-type-names
 - Stricter typing in Field Type Selector
 - Better names for field types
+
+## 0.0.1-dev-17-feature/DHP-25-example-page
+- Example survey.js format page added
 
 ## 0.0.1-dev-16-feature/DHP-10-visible-required
 - Add a selector to reference previous questions